Threaded proxy for token protected HTTP repos
Introduce a threaded internal proxy for remote package repositories that
require token based authentication.

This kind of authentication is not supported by package managers such as
dnf/yum. As soon as a repository is declared with `auth: idp_token` in
project's configuration, Rift starts a thread with an HTTP server that
acts as a proxy which injects the token in all HTTP requests to remote
repos.

The token is either extracted from RIFT_AUTH_IDP_TOKEN environment
variable or rift auth state file.

The URL to authenticated remote repositories are modified in Mock config
and in testing VM to redirect to Rift's internal proxy. The host IP
addresse in testing guest VM is systematically set in no_proxy
environment variable to avoid libcurl (used by dnf and many others)
sending the requests targeted Rift's internal proxy to IT global proxy.

+++ lib/rift/Mock.py
@@ -47,6 +47,7 @@ from rift.TempDir import TempDir
 from rift.RPM import RPM
 from rift.run import run_command
 from rift.Config import _DEFAULT_VARIANT
+from rift.proxy import AuthenticatedRepositoryProxyRuntime
 
 
 class Mock():
@@ -64,6 +65,7 @@ class Mock():
         self._config = config
         self._arch = arch
         self._tmpdir = None
+        self._repo_proxy = None
         self._mockname = f"rift-{self._arch}-{getpass.getuser()}"
         if proj_vers:
             self._mockname = f"{self._mockname}-{proj_vers}"
@@ -77,11 +79,15 @@ class Mock():
         for idx, repo in enumerate(repolist, start=1):
             assert repo.url is not None
             prio = repo.priority or (prio - 1)
+            repo_url = repo.generic_url(self._arch)
+            if self._repo_proxy and repo.authenticated():
+                repo_url = self._repo_proxy.repo_url(repo, "127.0.0.1")
             repo_ctx = {
                 'name': repo.name or f"repo{idx}",
                 'priority': prio,
-                'url': repo.generic_url(self._arch),
+                'url': repo_url,
                 'variants': repo.variants,
+                'authenticated': repo.authenticated(),
                 }
             if repo.module_hotfixes:
                 repo_ctx['module_hotfixes'] = repo.module_hotfixes
@@ -189,6 +195,8 @@ class Mock():
 
         This should be cleaned with clean().
         """
+        self._repo_proxy = AuthenticatedRepositoryProxyRuntime(self._config, repolist)
+        self._repo_proxy.start()
         self._init_tmp_conf(repolist)
         self._exec(['--init'])
 
@@ -235,7 +243,9 @@ class Mock():
 
     def clean(self):
         """Clean temporary files and RPMS created for this instance."""
-        self._tmpdir.delete()
+        if self._tmpdir:
+            self._tmpdir.delete()
+        self._stop_repo_proxy()
         for rpm in self.resultrpms():
             os.unlink(rpm.filepath)
 
@@ -243,6 +253,13 @@ class Mock():
         """Remove Mock environments (ie. chroots directories) from disk."""
         self._init_tmp_conf()
         self._exec(['--scrub=all'])
+
+    def _stop_repo_proxy(self):
+        """Stop repository proxy runtime if currently active."""
+        if self._repo_proxy is None:
+            return
+        self._repo_proxy.stop()
+        self._repo_proxy = None
 
     def build_srpm(self, specpath, sourcedir, sign):
         """
